requirements.txt
langchain
python-dotenv
langchain_groq
langchain_community
python-telegram-bot

#venv\Scripts\activate to activate the virtual environment.
#pip install -r requirements.txt to install the required packages.